fix init.sh for processes group
if a compiler was found, but compilation failed, the mission was
cancelled rather than falling back to using the shell scripts.

+++ missions/processes/01_ps_kill/init.sh
@@ -1,7 +1,6 @@
 #!/bin/sh
 
-_mission_init() {
-
+_compile() (
   if command -v gcc >/dev/null
   then
     CC=gcc
@@ -22,17 +21,11 @@ _mission_init() {
   elif command -v cc >/dev/null
   then
     CC=cc
-  elif ! [ -e "$MISSION_DIR/deps.sh" ] || ! command -v my_ps >/dev/null
-  then
-    # FIXME
-    echo "missing dummy mission!"
+  else
     return 1
   fi
-  mission_source "$MISSION_DIR/deps.sh" || return 1
 
-  if [ -n "$CC" ]
-  then
-    (
+  (
       # in debug mode, don't hide messages
       if [ "$GSH_MODE" != DEBUG ] || [ -z "$GSH_VERBOSE_DEBUG" ]
       then
@@ -40,19 +33,36 @@ _mission_init() {
         exec 2>/dev/null
       fi
       $CC $CFLAGS "$MISSION_DIR/spell.c" -o "$GSH_TMP/$(gettext "spell")"
-    ) || { echo "compilation failed" >&2; return 1; }
-  else
-    cp "$MISSION_DIR/spell.sh" "$GSH_TMP/$(gettext "spell")"
-    chmod 755 "$GSH_TMP/$(gettext "spell")"
+  ) || { echo "compilation failed" >&2; return 1; }
+  return 0
+)
+
+_install_script() (
+  if ! [ -e "$MISSION_DIR/deps.sh" ] || ! command -v my_ps >/dev/null
+  then
+    # FIXME
+    echo "missing dummy mission!"
+    return 1
   fi
+  mission_source "$MISSION_DIR/deps.sh" || return 1
+
+  cp "$MISSION_DIR/spell.sh" "$GSH_TMP/$(gettext "spell")"
+  chmod 755 "$GSH_TMP/$(gettext "spell")"
+)
+
+if _compile || _install_script
+then
+  unset -f _compile _install_script
+
+  [ -n "$GSH_NON_INTERACTIVE" ] || set +o monitor  # do not monitor background processes
+  # FIXME: for some unknown reason, this doesn't work if we start with this
+  # mission directly!
+
+  # set +b 2>/dev/null    # POSIX, but not supported by zsh
+
   "$GSH_TMP/$(gettext "spell")" &
   echo $! > "$GSH_TMP"/spell.pid
-  return 0
-}
-
-[ -n "$GSH_NON_INTERACTIVE" ] || set +o monitor  # do not monitor background processes
-# FIXME: for some unknown reason, this doesn't work if we start with this
-# mission directly!
-
-# set +b 2>/dev/null    # POSIX, but not supported by zsh
-_mission_init
+else
+  unset -f _compile _install_script
+  false
+fi
